@@ -1,23 +1,40 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-import { NestFactory } from '@nestjs/core';
+import { NestFactory, Reflector } from '@nestjs/core';
+import { ClassSerializerInterceptor, ValidationPipe } from '@nestjs/common';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 import { AppModule } from './app.module';
-
-const version = require('../package.json').version;
+import { HttpExceptionFilter } from './common/filters/http-exception.filter';
+import { LoggingInterceptor } from './common/interceptors/logging.interceptor';
+import helmet from 'helmet';
+import { version } from '../package.json';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
 
-  app.enableCors({
-    origin: '*',
-  });
+  const allowedOrigins = process.env.ALLOWED_ORIGINS
+    ? process.env.ALLOWED_ORIGINS.split(',')
+    : '*';
+  app.enableCors({ origin: allowedOrigins });
+  app.use(helmet());
+
+  app.useGlobalPipes(
+    new ValidationPipe({
+      whitelist: true,
+      forbidNonWhitelisted: true,
+      transform: true,
+      transformOptions: { enableImplicitConversion: true },
+    }),
+  );
+  app.useGlobalFilters(new HttpExceptionFilter());
+  app.useGlobalInterceptors(
+    new ClassSerializerInterceptor(app.get(Reflector)),
+    new LoggingInterceptor(),
+  );
 
   const options = new DocumentBuilder()
-    .setTitle(`💈App Barbearia`)
+    .setTitle(`💈 App Barbearia`)
     .setDescription('Back-end do projeto App Barbearia')
     .setVersion(version)
     .addBearerAuth()
-    .addTag('Endpoints:')
     .build();
 
   const document = SwaggerModule.createDocument(app, options);
@@ -26,7 +43,7 @@ async function bootstrap() {
 
   await app.listen(PORT, '0.0.0.0');
   console.log(
-    `💈app barbearia is in ${process.env.NODE_ENV} mode and is listening on port:`,
+    `💈 app barbearia is in ${process.env.NODE_ENV} mode and is listening on port:`,
     PORT,
   );
 }
