@@ -1,18 +1,22 @@
 import { BarberShopModule } from './barber-shop/barber-shop.module';
 import { Module } from '@nestjs/common';
+import { APP_GUARD } from '@nestjs/core';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { config } from 'dotenv';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { ThrottlerModule, ThrottlerGuard } from '@nestjs/throttler';
 import { UserModule } from './user/user.module';
 import { AuthModule } from './auth/auth.module';
 import { ServicesModule } from './service/service.module';
 import { SchedulingModule } from './scheduling/scheduling.module';
 import { BarberModule } from './barber/barber.module';
-
+import { AuditModule } from './common/audit/audit.module';
+import { HealthModule } from './health/health.module';
 config();
 
 @Module({
   imports: [
+    ThrottlerModule.forRoot([{ ttl: 60000, limit: 60 }]),
     TypeOrmModule.forRootAsync({
       imports: [ConfigModule],
       inject: [ConfigService],
@@ -33,6 +37,14 @@ config();
     ServicesModule,
     SchedulingModule,
     BarberModule,
+    AuditModule,
+    HealthModule,
+  ],
+  providers: [
+    {
+      provide: APP_GUARD,
+      useClass: ThrottlerGuard,
+    },
   ],
 })
 export class AppModule {}
