@@ -1,9 +1,21 @@
-import { Controller, Post, Body, Patch } from '@nestjs/common';
+import {
+  Controller,
+  Post,
+  Body,
+  Patch,
+  UseGuards,
+  Get,
+  Req,
+} from '@nestjs/common';
+import { Throttle, SkipThrottle } from '@nestjs/throttler';
+import { AuthGuard } from '@nestjs/passport';
 import { AuthService } from './auth.service';
 import {
   ApiTags,
   ApiOperation,
+  ApiBearerAuth,
   ApiForbiddenResponse,
+  ApiOkResponse,
   ApiUnauthorizedResponse,
   ApiNotFoundResponse,
 } from '@nestjs/swagger';
@@ -19,10 +31,12 @@ export class AuthController {
     private userService: UserService,
   ) {}
 
+  @Throttle({ default: { ttl: 60000, limit: 5 } })
   @ApiOperation({
     summary: 'Autentica o usuário',
-    description: `Roles: ${process.env.ALL}`,
+    description: 'Retorna o token JWT para acesso aos endpoints protegidos.',
   })
+  @ApiOkResponse({ description: 'Token JWT retornado' })
   @ApiUnauthorizedResponse({
     description: 'Senha inválida',
   })
@@ -34,11 +48,30 @@ export class AuthController {
     return await this.authService.validateUserByPassword(auth);
   }
 
+  @SkipThrottle()
+  @UseGuards(AuthGuard())
+  @ApiBearerAuth()
+  @ApiOperation({
+    summary: 'Retorna o perfil do usuário autenticado',
+    description: 'Usa o userId extraído do token JWT.',
+  })
+  @ApiOkResponse({ description: 'Perfil do usuário retornado' })
+  @ApiUnauthorizedResponse({ description: 'Token inválido ou ausente' })
+  @Get('/me')
+  async getMe(@Req() req: any) {
+    return await this.authService.getMe(req.user.userId);
+  }
+
+  @SkipThrottle()
+  @UseGuards(AuthGuard())
+  @ApiBearerAuth()
   @ApiOperation({
     summary: 'Recuperação de senha do usuário',
     description:
       'Permite que o usuário recupere sua senha usando a data de nascimento e o documento.',
   })
+  @ApiOkResponse({ description: 'Senha alterada com sucesso' })
+  @ApiUnauthorizedResponse({ description: 'Token inválido ou ausente' })
   @ApiNotFoundResponse({
     description: 'Usuário não encontrado',
   })
