@@ -5,8 +5,8 @@ export const mockBarberShop = {
   name: 'Barber Shop Name',
   document: '12.345.678/0001-99',
   address: '123 Main St',
-  lat: '-19.917299',
-  long: '-43.934559',
+  lat: -19.917299,
+  long: -43.934559,
   cellphone: '31985171031',
   email: 'barbershop@example.com',
 } as BarberShop;
