@@ -1,26 +1,32 @@
 import { MigrationInterface, QueryRunner } from 'typeorm';
 
-export class Migration1724102526544 implements MigrationInterface {
-  name = 'Migration1724102526544';
+export class Migration1771952395805 implements MigrationInterface {
+  name = 'Migration1771952395805';
 
   public async up(queryRunner: QueryRunner): Promise<void> {
     await queryRunner.query(
       `CREATE TYPE "public"."service_type_enum" AS ENUM('hair', 'beard', 'eyebrow', 'skincare', 'other')`,
     );
     await queryRunner.query(
-      `CREATE TABLE "service" ("id" uuid NOT NULL DEFAULT uuid_generate_v4(), "createdAt" TIMESTAMP NOT NULL DEFAULT now(), "updatedAt" TIMESTAMP NOT NULL DEFAULT now(), "active" boolean NOT NULL DEFAULT true, "name" character varying, "type" "public"."service_type_enum" NOT NULL DEFAULT 'hair', "value" character varying, CONSTRAINT "PK_85a21558c006647cd76fdce044b" PRIMARY KEY ("id"))`,
+      `CREATE TABLE "service" ("id" uuid NOT NULL DEFAULT uuid_generate_v4(), "createdAt" TIMESTAMP NOT NULL DEFAULT now(), "updatedAt" TIMESTAMP NOT NULL DEFAULT now(), "active" boolean NOT NULL DEFAULT true, "deletedAt" TIMESTAMP, "name" character varying, "type" "public"."service_type_enum" NOT NULL DEFAULT 'hair', "price" numeric(10,2), "durationMinutes" integer, CONSTRAINT "PK_85a21558c006647cd76fdce044b" PRIMARY KEY ("id"))`,
     );
     await queryRunner.query(
-      `CREATE TABLE "barber" ("id" uuid NOT NULL DEFAULT uuid_generate_v4(), "createdAt" TIMESTAMP NOT NULL DEFAULT now(), "updatedAt" TIMESTAMP NOT NULL DEFAULT now(), "active" boolean NOT NULL DEFAULT true, "name" character varying, "document" character varying(11), "email" character varying, "cellphone" character varying, CONSTRAINT "PK_393a066f1a87c8642e776ba7054" PRIMARY KEY ("id"))`,
+      `CREATE TABLE "barber" ("id" uuid NOT NULL DEFAULT uuid_generate_v4(), "createdAt" TIMESTAMP NOT NULL DEFAULT now(), "updatedAt" TIMESTAMP NOT NULL DEFAULT now(), "active" boolean NOT NULL DEFAULT true, "deletedAt" TIMESTAMP, "name" character varying, "document" character varying(11), "email" character varying, "cellphone" character varying, CONSTRAINT "UQ_d923108cd4a381f456f99f8dad1" UNIQUE ("email", "document"), CONSTRAINT "PK_393a066f1a87c8642e776ba7054" PRIMARY KEY ("id"))`,
     );
     await queryRunner.query(
-      `CREATE TABLE "scheduling" ("id" uuid NOT NULL DEFAULT uuid_generate_v4(), "createdAt" TIMESTAMP NOT NULL DEFAULT now(), "updatedAt" TIMESTAMP NOT NULL DEFAULT now(), "active" boolean NOT NULL DEFAULT true, "date" TIMESTAMP, "userId" uuid, "barbershopId" uuid, "barberId" uuid, CONSTRAINT "PK_a19510fdc2c3f1c9daff8b6e395" PRIMARY KEY ("id"))`,
+      `CREATE TYPE "public"."scheduling_status_enum" AS ENUM('pending', 'confirmed', 'completed', 'cancelled', 'no_show')`,
     );
     await queryRunner.query(
-      `CREATE TABLE "barber_shop" ("id" uuid NOT NULL DEFAULT uuid_generate_v4(), "createdAt" TIMESTAMP NOT NULL DEFAULT now(), "updatedAt" TIMESTAMP NOT NULL DEFAULT now(), "active" boolean NOT NULL DEFAULT true, "name" character varying, "document" character varying, "address" character varying, "lat" character varying, "long" character varying, "cellphone" character varying, "email" character varying, CONSTRAINT "PK_1f886aa2348269079caccdd1ad4" PRIMARY KEY ("id"))`,
+      `CREATE TABLE "scheduling" ("id" uuid NOT NULL DEFAULT uuid_generate_v4(), "createdAt" TIMESTAMP NOT NULL DEFAULT now(), "updatedAt" TIMESTAMP NOT NULL DEFAULT now(), "active" boolean NOT NULL DEFAULT true, "deletedAt" TIMESTAMP, "date" TIMESTAMP, "endTime" TIMESTAMP, "status" "public"."scheduling_status_enum" NOT NULL DEFAULT 'pending', "userId" uuid, "barbershopId" uuid, "barberId" uuid, CONSTRAINT "PK_a19510fdc2c3f1c9daff8b6e395" PRIMARY KEY ("id"))`,
     );
     await queryRunner.query(
-      `CREATE TABLE "user" ("id" uuid NOT NULL DEFAULT uuid_generate_v4(), "createdAt" TIMESTAMP NOT NULL DEFAULT now(), "updatedAt" TIMESTAMP NOT NULL DEFAULT now(), "active" boolean NOT NULL DEFAULT true, "name" character varying, "birthdate" TIMESTAMP, "document" character varying, "email" character varying, "password" character varying, "cellphone" character varying(20), "userType" character varying NOT NULL DEFAULT 'user', CONSTRAINT "UQ_335a7b396d5a7b840e2e31200b4" UNIQUE ("email", "document"), CONSTRAINT "PK_cace4a159ff9f2512dd42373760" PRIMARY KEY ("id"))`,
+      `CREATE TABLE "barber_shop" ("id" uuid NOT NULL DEFAULT uuid_generate_v4(), "createdAt" TIMESTAMP NOT NULL DEFAULT now(), "updatedAt" TIMESTAMP NOT NULL DEFAULT now(), "active" boolean NOT NULL DEFAULT true, "deletedAt" TIMESTAMP, "name" character varying, "document" character varying, "address" character varying, "lat" numeric(10,7), "long" numeric(10,7), "cellphone" character varying, "email" character varying, CONSTRAINT "PK_1f886aa2348269079caccdd1ad4" PRIMARY KEY ("id"))`,
+    );
+    await queryRunner.query(
+      `CREATE TABLE "user" ("id" uuid NOT NULL DEFAULT uuid_generate_v4(), "createdAt" TIMESTAMP NOT NULL DEFAULT now(), "updatedAt" TIMESTAMP NOT NULL DEFAULT now(), "active" boolean NOT NULL DEFAULT true, "deletedAt" TIMESTAMP, "name" character varying, "birthdate" TIMESTAMP, "document" character varying, "email" character varying, "password" character varying, "cellphone" character varying(20), "userType" character varying NOT NULL DEFAULT 'user', CONSTRAINT "UQ_335a7b396d5a7b840e2e31200b4" UNIQUE ("email", "document"), CONSTRAINT "PK_cace4a159ff9f2512dd42373760" PRIMARY KEY ("id"))`,
+    );
+    await queryRunner.query(
+      `CREATE TABLE "barbershop_business_hours" ("id" uuid NOT NULL DEFAULT uuid_generate_v4(), "createdAt" TIMESTAMP NOT NULL DEFAULT now(), "updatedAt" TIMESTAMP NOT NULL DEFAULT now(), "active" boolean NOT NULL DEFAULT true, "deletedAt" TIMESTAMP, "dayOfWeek" integer NOT NULL, "openTime" character varying(5) NOT NULL, "closeTime" character varying(5) NOT NULL, "isOpen" boolean NOT NULL DEFAULT true, "barbershopId" uuid, CONSTRAINT "PK_d24090230f118e7ba5a6451395f" PRIMARY KEY ("id"))`,
     );
     await queryRunner.query(
       `CREATE TABLE "barbershop_services" ("serviceId" uuid NOT NULL, "barberShopId" uuid NOT NULL, CONSTRAINT "PK_918d7c01233c7915484e23d354a" PRIMARY KEY ("serviceId", "barberShopId"))`,
@@ -57,6 +63,9 @@ export class Migration1724102526544 implements MigrationInterface {
     );
     await queryRunner.query(
       `ALTER TABLE "scheduling" ADD CONSTRAINT "FK_843feaec9e559db38215f0212c9" FOREIGN KEY ("barberId") REFERENCES "barber"("id") ON DELETE NO ACTION ON UPDATE NO ACTION`,
+    );
+    await queryRunner.query(
+      `ALTER TABLE "barbershop_business_hours" ADD CONSTRAINT "FK_a64f73e99e0c7a83234c2f637b7" FOREIGN KEY ("barbershopId") REFERENCES "barber_shop"("id") ON DELETE CASCADE ON UPDATE NO ACTION`,
     );
     await queryRunner.query(
       `ALTER TABLE "barbershop_services" ADD CONSTRAINT "FK_fb87c32c2da706981e4353b7a01" FOREIGN KEY ("serviceId") REFERENCES "service"("id") ON DELETE CASCADE ON UPDATE CASCADE`,
@@ -98,6 +107,9 @@ export class Migration1724102526544 implements MigrationInterface {
       `ALTER TABLE "barbershop_services" DROP CONSTRAINT "FK_fb87c32c2da706981e4353b7a01"`,
     );
     await queryRunner.query(
+      `ALTER TABLE "barbershop_business_hours" DROP CONSTRAINT "FK_a64f73e99e0c7a83234c2f637b7"`,
+    );
+    await queryRunner.query(
       `ALTER TABLE "scheduling" DROP CONSTRAINT "FK_843feaec9e559db38215f0212c9"`,
     );
     await queryRunner.query(
@@ -127,9 +139,11 @@ export class Migration1724102526544 implements MigrationInterface {
       `DROP INDEX "public"."IDX_fb87c32c2da706981e4353b7a0"`,
     );
     await queryRunner.query(`DROP TABLE "barbershop_services"`);
+    await queryRunner.query(`DROP TABLE "barbershop_business_hours"`);
     await queryRunner.query(`DROP TABLE "user"`);
     await queryRunner.query(`DROP TABLE "barber_shop"`);
     await queryRunner.query(`DROP TABLE "scheduling"`);
+    await queryRunner.query(`DROP TYPE "public"."scheduling_status_enum"`);
     await queryRunner.query(`DROP TABLE "barber"`);
     await queryRunner.query(`DROP TABLE "service"`);
     await queryRunner.query(`DROP TYPE "public"."service_type_enum"`);
